+++ tests/typing_examples.py
@@ -213,3 +213,6 @@ structlog.configure(
     wrapper_class=structlog.stdlib.BoundLogger,
     cache_logger_on_first_use=True,
 )
+
+with structlog.threadlocal.bound_threadlocal(x=42):
+    pass